Improvements: `uint8_t` is not support on some platforms (like TI C28x DSP), use `int` instead. And interface header has been moved into `include` folder.

+++ src/fmt_parser.c
@@ -1,12 +1,11 @@
 #include "fmt_parser.h"
 #include <stddef.h>
-#include <stdint.h>
 
 static int cmp_n(const char *s1, const char *s2, int size);
 
 struct placeholder_table_item {
 	const char *name;
-	uint8_t size;
+	int size;
 	enum fmt_parser_data_t type;
 };
